testframework/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
crontab==0.22.6
cycler==0.10.0
idna==2.8
kiwisolver==1.1.0
numpy==1.16.4
pyparsing==2.4.0
PyQt5-sip==4.19.17
pystache==0.5.4
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
selenium==3.141.0
simplejson==3.16.0
six==1.12.0
urllib3==1.25.3
WMI==1.4.9
raven==6.10.0
apscheduler==3.6.1